requirements.txt
## Runtime dependencies (pinned with conservative upper bounds for stability)
# FastMCP server framework
fastmcp>=0.1.0,<0.2.0
# HTTP requests to public + G2B endpoints
requests>=2.31.0,<2.33.0
# Async HTTP client
aiohttp>=3.9.0,<3.10.0
# .env secret loading (NARAMARKET_SERVICE_KEY)
python-dotenv>=1.0.0,<1.1.0
# DataFrame + CSV/Parquet handling
pandas>=2.2.0,<2.3.0
# Parquet writer/reader backend
pyarrow>=15.0.0,<18.0.0
# FastAPI and HTTP server
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
# Authentication and security
PyJWT>=2.8.0,<2.9.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.1.0

## Development dependencies (optional)
# pytest>=8.0.0,<9.0.0
# mypy>=1.10.0,<1.11.0
# types-requests>=2.31.0,<2.32.0
# httpx>=0.25.0,<0.26.0  # for async testing
